docs: improved instructions for samples and tests (#41)
