+++ dialogflow-cx/list-intents.js
@@ -25,6 +25,11 @@ async function main(projectId, location, agentId) {
 
   // Imports the Google Cloud Some API library
   const {IntentsClient} = require('@google-cloud/dialogflow-cx');
+  /**
+   * Example for regional endpoint:
+   *   const location = 'us-central1'
+   *   const client = new SessionsClient({apiEndpoint: 'us-central1-dialogflow.googleapis.com'})
+   */
   const client = new IntentsClient();
 
   async function listIntents() {
