+++ dialogflow-cx/quickstart.js
@@ -37,6 +37,11 @@ async function main(
 
   // Imports the Google Cloud Some API library
   const {SessionsClient} = require('@google-cloud/dialogflow-cx');
+  /**
+   * Example for regional endpoint:
+   *   const location = 'us-central1'
+   *   const client = new SessionsClient({apiEndpoint: 'us-central1-dialogflow.googleapis.com'})
+   */
   const client = new SessionsClient();
 
   const fs = require('fs');
